feat(settings/avatar): page.tsx RSC 변환 + 모바일 가드
기존 'use client' 본문을 AvatarSettingsPageDesktop 위젯으로 추출 (chunk 1
(lobby)/page.tsx 와 동일 패턴). page.tsx 가 headers() 보고 모바일은 가드 카드,
데스크탑은 desktop shell 렌더.
