+++ src/widgets/avatar-settings-page-desktop/index.ts
@@ -1,0 +1,1 @@
+export { default as AvatarSettingsPageDesktop } from './avatar-settings-page-desktop.component';
